K8s/Test: Allow setting license path in test helper (#99786)
